requirements.txt
Django
environs==9.5.*
Pillow==8.4.*
django-admin-sortable2
django-tinymce
